[ADD] authentication page
