+++ android/app/src/main/res/drawable-night-xhdpi/android12splash.png
