+++ android/app/src/main/res/drawable-night-xxhdpi/android12splash.png
