+++ lib/constants/colors.dart
@@ -6,6 +6,7 @@ Color lightRedColor = const Color(0xFFF8C8C8);
 Color greenColor = const Color(0xFF388E3C);
 Color blackColor = const Color(0xFF000000);
 Color yellowColor = const Color(0xFFFFBC3A);
+Color dartYellowColor = const Color.fromARGB(255, 216, 158, 41);
 Color lightYellowColor = const Color(0xFFFFECB3);
 Color greyColor = const Color(0xFFD9D9D9);
 Color whiteColor = const Color(0xFFFFFFFF);
