+++ lib/services/auth/auth_service.dart
@@ -1,0 +1,17 @@
+import 'package:firebase_auth/firebase_auth.dart';
+
+class AuthService {
+  final FirebaseAuth _firebaseAuth = FirebaseAuth.instance;
+
+  //sign user in
+  Future<UserCredential> signInWithEmailandPassword(String email, String password) async{
+    try {
+      UserCredential userCredential = await _firebaseAuth.signInWithEmailAndPassword(email: email, password: password);
+      return userCredential;
+    } on FirebaseAuthException catch (e) {
+      throw  Exception(e.code);
+    }
+  }
+
+  //sign user out
+}
