+++ lib/utilities/drawer.dart
@@ -60,6 +60,54 @@ class BakingUpDrawer extends StatelessWidget {
             },
           ),
           ListTile(
+            tileColor: currentDrawerIndex == 1
+                ? const Color.fromRGBO(217, 217, 217, 0.5)
+                : Colors.white,
+            title: const Row(
+              children: [
+                Text(
+                  'vertify',
+                  style: TextStyle(
+                    color: Colors.black,
+                    fontFamily: 'Poppins',
+                    fontStyle: FontStyle.normal,
+                    fontWeight: FontWeight.w400,
+                    fontSize: 10,
+                    height: 0.5,
+                  ),
+                ),
+              ],
+            ),
+            onTap: () {
+              Navigator.of(context)
+                  .pushNamedAndRemoveUntil(VerifyEmailRoute, (route) => false);
+            },
+          ),
+          ListTile(
+            tileColor: currentDrawerIndex == 1
+                ? const Color.fromRGBO(217, 217, 217, 0.5)
+                : Colors.white,
+            title: const Row(
+              children: [
+                Text(
+                  'forget password',
+                  style: TextStyle(
+                    color: Colors.black,
+                    fontFamily: 'Poppins',
+                    fontStyle: FontStyle.normal,
+                    fontWeight: FontWeight.w400,
+                    fontSize: 10,
+                    height: 0.5,
+                  ),
+                ),
+              ],
+            ),
+            onTap: () {
+              Navigator.of(context)
+                  .pushNamedAndRemoveUntil(forgotPasswordRoute, (route) => false);
+            },
+          ),
+          ListTile(
             tileColor: currentDrawerIndex == 2
                 ? const Color.fromRGBO(217, 217, 217, 0.5)
                 : Colors.white,
