+++ lib/widgets/auth/confirm_password_text_form_field.dart
@@ -1,5 +1,6 @@
 import 'package:bakingup_frontend/constants/colors.dart';
 import 'package:flutter/material.dart';
+
 class ConfirmPasswordFormField extends StatefulWidget {
   final TextEditingController passwordController;
   final TextEditingController confirmPasswordController;
@@ -11,7 +12,7 @@ class ConfirmPasswordFormField extends StatefulWidget {
     required this.passwordController,
     required this.hintText,
     required this.confirmPasswordController,
-    required this.error
+    required this.error,
   });
 
   @override
@@ -20,61 +21,99 @@ class ConfirmPasswordFormField extends StatefulWidget {
 
 class _ConfirmPasswordFormFieldState extends State<ConfirmPasswordFormField> {
   bool passwordVisible = true;
-  ValueNotifier<bool> passwordMatchNotifier = ValueNotifier<bool>(false);
+  bool hasError = false;
+  bool isFocused = false;
 
   @override
   Widget build(BuildContext context) {
-    return ValueListenableBuilder<bool>(
-      valueListenable: passwordMatchNotifier,
-      builder: (context, passwordMatch, _) {
-        final borderColor = passwordMatch ? Colors.green : Colors.red;
-        return Builder(
-          builder: (BuildContext context) {
-            return TextFormField(
-              validator: (value){
-                if (value == null ||
-                    value.isEmpty ) {
-                  widget.error();
-                  return 'Please enter password';
-                }else if(value != widget.passwordController.text){
-                  widget.error();
-                    return "Password must be same as above";
-                }
-                return null;
-              },
-              obscureText: passwordVisible,
-              controller: widget.confirmPasswordController,
-              decoration: InputDecoration(
-                focusedBorder: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(25.0),
-                  borderSide: BorderSide(
-                    color: borderColor,
-                    width: 1.5
-                  ),
-                ),
-                border: const OutlineInputBorder(
-                  borderRadius: BorderRadius.all(Radius.circular(25)),
-                ),
-                hintText: widget.hintText,
-                contentPadding: const EdgeInsets.symmetric(horizontal: 25, vertical: 15),
-                suffixIcon: IconButton(
-                  onPressed: () {
-                    setState(() {
-                      passwordVisible = !passwordVisible;
-                    });
-                  },
-                  icon: Icon(passwordVisible ? Icons.visibility : Icons.visibility_off, color: darkBeigeColor,),
+    return Focus(
+      onFocusChange: (focus) {
+        setState(() {
+          isFocused = focus;
+        });
+      },
+      child: ValueListenableBuilder<TextEditingValue>(
+        valueListenable: widget.confirmPasswordController,
+        builder: (context, value, child) {
+          final confirmPassword = value.text;
+          final password = widget.passwordController.text;
+
+          bool passwordsMatch = confirmPassword == password && confirmPassword.isNotEmpty;
+
+          Color borderColor;
+
+          if (hasError) {
+            borderColor = errorRedColor; 
+          } else if (isFocused && passwordsMatch) {
+            borderColor = greenColor; 
+          } else if (isFocused && confirmPassword.isEmpty) {
+            borderColor = darkBrownColor; 
+          } else if (!passwordsMatch && !isFocused && confirmPassword.isNotEmpty) {
+            borderColor = errorRedColor; 
+          } else {
+            borderColor = darkBrownColor; 
+          }
+
+          return TextFormField(
+            validator: (value) {
+              if (value == null || value.isEmpty) {
+                widget.error();
+                setState(() {
+                  hasError = true;
+                });
+                return 'Please enter password';
+              } else if (value != password) {
+                widget.error();
+                setState(() {
+                  hasError = true;
+                });
+                return "Password must be the same as above";
+              } else {
+                setState(() {
+                  hasError = false;
+                });
+              }
+              return null;
+            },
+            obscureText: passwordVisible,
+            controller: widget.confirmPasswordController,
+            decoration: InputDecoration(
+              focusedBorder: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(10),
+                borderSide: BorderSide(
+                  color: borderColor,
+                  width: 2,
                 ),
               ),
-              onChanged: (value) {
-                final password = widget.passwordController.text;
-                final confirmPassword = value;
-                passwordMatchNotifier.value = password == confirmPassword;
-              },
-            );
-          },
-        );
-      },
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(10),
+                // borderSide: BorderSide(
+                //   color: borderColor,
+                //   width: 1.5,
+                // ),
+              ),
+              hintText: widget.hintText,
+              contentPadding: const EdgeInsets.symmetric(horizontal: 25, vertical: 15),
+              suffixIcon: IconButton(
+                onPressed: () {
+                  setState(() {
+                    passwordVisible = !passwordVisible;
+                  });
+                },
+                icon: Icon(
+                  passwordVisible ? Icons.visibility : Icons.visibility_off,
+                  color: darkBeigeColor,
+                ),
+              ),
+            ),
+            onChanged: (value) {
+              setState(() {
+                hasError = value != widget.passwordController.text && value.isNotEmpty;
+              });
+            },
+          );
+        },
+      ),
     );
   }
 }
