+++ lib/widgets/auth/register_text_form_field.dart
@@ -1,7 +1,12 @@
+import 'package:bakingup_frontend/constants/colors.dart';
 import 'package:flutter/material.dart';
 
 class RegisterTextFormField extends StatelessWidget {
-  const RegisterTextFormField({super.key, required this.validate, required this.hintText, required this.controller});
+  const RegisterTextFormField(
+      {super.key,
+      required this.validate,
+      required this.hintText,
+      required this.controller});
   final String? Function(String?)? validate;
   final String hintText;
   final TextEditingController controller;
@@ -13,11 +18,14 @@ class RegisterTextFormField extends StatelessWidget {
       validator: validate,
       decoration: InputDecoration(
         border: const OutlineInputBorder(
-            borderRadius:
-            BorderRadius.all(Radius.circular(25))),
+            borderRadius: BorderRadius.all(Radius.circular(10))),
         hintText: hintText,
-        contentPadding: const EdgeInsets.symmetric(
-            horizontal: 25, vertical: 15),
+        focusedBorder: OutlineInputBorder(
+          borderSide: BorderSide(color: darkBrownColor, width: 2),
+          borderRadius: BorderRadius.circular(10),
+        ),
+        contentPadding:
+            const EdgeInsets.symmetric(horizontal: 25, vertical: 15),
       ),
     );
   }
