+++ linux/flutter/generated_plugin_registrant.cc
@@ -6,6 +6,10 @@
 
 #include "generated_plugin_registrant.h"
 
+#include <flutter_platform_alert/flutter_platform_alert_plugin.h>
 
 void fl_register_plugins(FlPluginRegistry* registry) {
+  g_autoptr(FlPluginRegistrar) flutter_platform_alert_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "FlutterPlatformAlertPlugin");
+  flutter_platform_alert_plugin_register_with_registrar(flutter_platform_alert_registrar);
 }
